Clean up player and visualizer imports
