+++ music/demos/callback_player.ts
@@ -92,8 +92,8 @@ class MetronomeCallback extends mm.BasePlayerCallback {
       ]
     ];
     this.drumPitchToClass = new Map<number, number>();
-    for (let c = 0; c < mm.data.DEFAULT_DRUM_PITCH_CLASSES.length; ++c) {
-      mm.data.DEFAULT_DRUM_PITCH_CLASSES[c].forEach((p) => {
+    for (let c = 0; c < mm.constants.DEFAULT_DRUM_PITCH_CLASSES.length; ++c) {
+      mm.constants.DEFAULT_DRUM_PITCH_CLASSES[c].forEach((p) => {
         this.drumPitchToClass.set(p, c);
       });
     }
