+++ music/src/core/constants.ts
@@ -37,6 +37,26 @@ export const DEFAULT_CHANNEL = 0;
 export const DRUM_CHANNEL = 9;
 export const NON_DRUM_CHANNELS : ReadonlyArray<number> =
     [0, 1, 2, 3, 4, 5, 6, 7, 8, 10, 11, 12, 13, 14, 15];
+export const DEFAULT_DRUM_PITCH_CLASSES: number[][] = [
+  // bass drum
+  [36, 35],
+  // snare drum
+  [38, 27, 28, 31, 32, 33, 34, 37, 39, 40, 56, 65, 66, 75, 85],
+  // closed hi-hat
+  [42, 44, 54, 68, 69, 70, 71, 73, 78, 80],
+  // open hi-hat
+  [46, 67, 72, 74, 79, 81],
+  // low tom
+  [45, 29, 41, 61, 64, 84],
+  // mid tom
+  [48, 47, 60, 63, 77, 86, 87],
+  // high tom
+  [50, 30, 43, 62, 76, 83],
+  // crash cymbal
+  [49, 55, 57, 58],
+  // ride cymbal
+  [51, 52, 53, 59, 82]
+];
 
 // Velocity-related constants.
 export const MIN_MIDI_VELOCITY = 0;
